Fix bug introduced in revision 28895 and change the steps in examples to work with types::global_dof_index.
git-svn-id: https://svn.dealii.org/branches/branch_bigger_global_dof_indices_4@28903 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/source/base/utilities.cc
@@ -479,7 +479,7 @@ namespace Utilities
 
 
   std::vector<unsigned long long int>
-  invert_permutation (const std::vector<unsigned long int> &permutation)
+  invert_permutation (const std::vector<unsigned long long int> &permutation)
   {
     const unsigned long long int n = permutation.size();
